refactor(coderd/x/chatd): extract Anthropic tool sanitization
