+++ coderd/x/chatd/chatloop/chatloop.go
@@ -26,14 +26,14 @@ import (
 	"github.com/coder/coder/v2/coderd/x/chatd/chaterror"
 	"github.com/coder/coder/v2/coderd/x/chatd/chatprompt"
 	"github.com/coder/coder/v2/coderd/x/chatd/chatretry"
+	"github.com/coder/coder/v2/coderd/x/chatd/chatsanitize"
 	"github.com/coder/coder/v2/coderd/x/chatd/chattool"
 	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/quartz"
 )
 
 const (
-	maxAnthropicProviderToolViolationLogDetails = 32
-	interruptedToolResultErrorMessage           = "tool call was interrupted before it produced a result"
+	interruptedToolResultErrorMessage = "tool call was interrupted before it produced a result"
 	// maxCompactionRetries limits how many times the post-run
 	// compaction safety net can re-enter the step loop. This
 	// prevents infinite compaction loops when the model keeps
@@ -392,13 +392,13 @@ func Run(ctx context.Context, opts RunOptions) error {
 			}
 			prepared := make([]fantasy.Message, len(messages))
 			copy(prepared, messages)
-			prepared, sanitizeStats := chatprompt.SanitizeAnthropicProviderToolHistory(provider, prepared)
-			chatprompt.LogAnthropicProviderToolSanitization(
+			prepared, sanitizeStats := chatsanitize.SanitizeAnthropicProviderToolHistory(provider, prepared)
+			chatsanitize.LogAnthropicProviderToolSanitization(
 				ctx, opts.Logger, "pre_request", provider, modelName, sanitizeStats,
 				slog.F("step_index", step),
 				slog.F("total_steps", totalSteps),
 			)
-			prepared = applyAnthropicProviderToolGuard(
+			prepared = chatsanitize.ApplyAnthropicProviderToolGuard(
 				ctx, opts.Logger, provider, modelName, prepared,
 			)
 			if applyAnthropicCaching {
@@ -533,7 +533,7 @@ func Run(ctx context.Context, opts RunOptions) error {
 						opts.ContextLimitFallback,
 					)
 
-					result.content = sanitizeAnthropicProviderToolStepContent(
+					result.content = chatsanitize.SanitizeAnthropicProviderToolStepContent(
 						ctx, opts.Logger, provider, modelName,
 						"dynamic_tool_persist", step, result.finishReason, result.content,
 					)
@@ -580,7 +580,7 @@ func Run(ctx context.Context, opts RunOptions) error {
 				result.providerMetadata,
 				opts.ContextLimitFallback,
 			)
-			result.content = sanitizeAnthropicProviderToolStepContent(
+			result.content = chatsanitize.SanitizeAnthropicProviderToolStepContent(
 				ctx, opts.Logger, provider, modelName,
 				"normal_persist", step, result.finishReason, result.content,
 			)
@@ -736,675 +736,6 @@ func Run(ctx context.Context, opts RunOptions) error {
 	}
 
 	return nil
-}
-
-func sanitizeAnthropicProviderToolStepContent(
-	ctx context.Context,
-	logger slog.Logger,
-	provider string,
-	modelName string,
-	phase string,
-	step int,
-	finishReason fantasy.FinishReason,
-	content []fantasy.Content,
-) []fantasy.Content {
-	sanitized, stats := sanitizeAnthropicProviderToolContent(provider, content)
-	chatprompt.LogAnthropicProviderToolSanitization(
-		ctx, logger, phase, provider, modelName, stats,
-		slog.F("step_index", step),
-		slog.F("finish_reason", finishReason),
-	)
-	return sanitized
-}
-
-func sanitizeAnthropicProviderToolContent(
-	provider string,
-	content []fantasy.Content,
-) ([]fantasy.Content, chatprompt.AnthropicProviderToolSanitizationStats) {
-	var stats chatprompt.AnthropicProviderToolSanitizationStats
-	if provider != fantasyanthropic.Name || len(content) == 0 {
-		return content, stats
-	}
-
-	partIndexByContentIndex := make([]int, len(content))
-	for index := range partIndexByContentIndex {
-		partIndexByContentIndex[index] = noMappedToolPartIndex
-	}
-	contentKinds := make([]mappedToolContentKind, len(content))
-	parts := make([]fantasy.MessagePart, 0, len(content))
-	providerCalls := make(map[string][]mappedProviderToolCall)
-	providerResultNames := make(map[string][]string)
-	for contentIndex, block := range content {
-		if toolCall, ok := safeToolCallContent(block); ok {
-			partIndex := len(parts)
-			parts = append(parts, toolCallContentToPart(toolCall))
-			partIndexByContentIndex[contentIndex] = partIndex
-			contentKinds[contentIndex] = mappedToolContentCall
-			if toolCall.ProviderExecuted {
-				providerCalls[toolCall.ToolCallID] = append(
-					providerCalls[toolCall.ToolCallID],
-					mappedProviderToolCall{
-						partIndex: partIndex,
-						toolName:  toolCall.ToolName,
-					},
-				)
-			}
-			continue
-		}
-		if toolResult, ok := safeToolResultContent(block); ok {
-			partIndex := len(parts)
-			parts = append(parts, toolResultContentToPart(toolResult))
-			partIndexByContentIndex[contentIndex] = partIndex
-			contentKinds[contentIndex] = mappedToolContentResult
-			if toolResult.ProviderExecuted {
-				providerResultNames[toolResult.ToolCallID] = append(
-					providerResultNames[toolResult.ToolCallID],
-					toolResult.ToolName,
-				)
-			}
-		}
-	}
-	if len(parts) == 0 {
-		return content, stats
-	}
-
-	// ToolResultContent carries ToolName, but ToolResultPart does not. Preserve
-	// the content sanitizer mismatch check by invalidating the synthetic call.
-	for id, calls := range providerCalls {
-		for _, call := range calls {
-			for _, resultToolName := range providerResultNames[id] {
-				if resultToolName == "" || resultToolName == call.toolName {
-					continue
-				}
-				toolCall, ok := parts[call.partIndex].(fantasy.ToolCallPart)
-				if !ok {
-					break
-				}
-				toolCall.ToolName = ""
-				parts[call.partIndex] = toolCall
-				break
-			}
-		}
-	}
-
-	removeParts, _ := chatprompt.AnthropicProviderToolPartsToRemove(provider, parts)
-	if len(removeParts) == 0 {
-		return content, stats
-	}
-
-	removeContent := make(map[int]struct{}, len(removeParts))
-	for contentIndex, partIndex := range partIndexByContentIndex {
-		if partIndex == noMappedToolPartIndex {
-			continue
-		}
-		if _, remove := removeParts[partIndex]; remove {
-			removeContent[contentIndex] = struct{}{}
-		}
-	}
-	if len(removeContent) == 0 {
-		return content, stats
-	}
-
-	out := make([]fantasy.Content, 0, len(content))
-	for contentIndex, block := range content {
-		if _, remove := removeContent[contentIndex]; remove {
-			switch contentKinds[contentIndex] {
-			case mappedToolContentCall:
-				stats.RemovedToolCalls++
-			case mappedToolContentResult:
-				stats.RemovedToolResults++
-				if textContent, ok := anthropicProviderToolResultTextContent(block); ok {
-					out = append(out, textContent)
-				}
-			}
-			continue
-		}
-		out = append(out, block)
-	}
-	return out, stats
-}
-
-const noMappedToolPartIndex = -1
-
-type mappedToolContentKind int
-
-const (
-	_ mappedToolContentKind = iota
-	mappedToolContentCall
-	mappedToolContentResult
-)
-
-type mappedProviderToolCall struct {
-	partIndex int
-	toolName  string
-}
-
-func isAnthropicProviderExecutedToolCall(
-	provider string,
-	toolCall fantasy.ToolCallContent,
-) bool {
-	return provider == fantasyanthropic.Name && toolCall.ProviderExecuted
-}
-
-func anthropicProviderToolResultTextContent(
-	block fantasy.Content,
-) (fantasy.TextContent, bool) {
-	var zero fantasy.TextContent
-	toolResult, ok := safeToolResultContent(block)
-	if !ok || !toolResult.ProviderExecuted {
-		return zero, false
-	}
-	text := chatprompt.AnthropicToolResultOutputText(toolResult.Result)
-	if text == "" {
-		return zero, false
-	}
-	return fantasy.TextContent{Text: text}, true
-}
-
-func safeToolCallContent(block fantasy.Content) (fantasy.ToolCallContent, bool) {
-	var zero fantasy.ToolCallContent
-	switch value := block.(type) {
-	case fantasy.ToolCallContent:
-		return value, true
-	case *fantasy.ToolCallContent:
-		if value == nil {
-			return zero, false
-		}
-		return *value, true
-	default:
-		return zero, false
-	}
-}
-
-func safeToolResultContent(block fantasy.Content) (fantasy.ToolResultContent, bool) {
-	var zero fantasy.ToolResultContent
-	switch value := block.(type) {
-	case fantasy.ToolResultContent:
-		return value, true
-	case *fantasy.ToolResultContent:
-		if value == nil {
-			return zero, false
-		}
-		return *value, true
-	default:
-		return zero, false
-	}
-}
-
-func safeToolCallPart(part fantasy.MessagePart) (fantasy.ToolCallPart, bool) {
-	var zero fantasy.ToolCallPart
-	switch value := part.(type) {
-	case fantasy.ToolCallPart:
-		return value, true
-	case *fantasy.ToolCallPart:
-		if value == nil {
-			return zero, false
-		}
-		return *value, true
-	default:
-		return zero, false
-	}
-}
-
-func safeToolResultPart(part fantasy.MessagePart) (fantasy.ToolResultPart, bool) {
-	var zero fantasy.ToolResultPart
-	switch value := part.(type) {
-	case fantasy.ToolResultPart:
-		return value, true
-	case *fantasy.ToolResultPart:
-		if value == nil {
-			return zero, false
-		}
-		return *value, true
-	default:
-		return zero, false
-	}
-}
-
-func toolCallContentToPart(toolCall fantasy.ToolCallContent) fantasy.ToolCallPart {
-	return fantasy.ToolCallPart{
-		ToolCallID:       toolCall.ToolCallID,
-		ToolName:         toolCall.ToolName,
-		Input:            toolCall.Input,
-		ProviderExecuted: toolCall.ProviderExecuted,
-		ProviderOptions:  fantasy.ProviderOptions(toolCall.ProviderMetadata),
-	}
-}
-
-func toolResultContentToPart(toolResult fantasy.ToolResultContent) fantasy.ToolResultPart {
-	return fantasy.ToolResultPart{
-		ToolCallID:       toolResult.ToolCallID,
-		Output:           toolResult.Result,
-		ProviderExecuted: toolResult.ProviderExecuted,
-		ProviderOptions:  fantasy.ProviderOptions(toolResult.ProviderMetadata),
-	}
-}
-
-func applyAnthropicProviderToolGuard(
-	ctx context.Context,
-	logger slog.Logger,
-	provider string,
-	modelName string,
-	messages []fantasy.Message,
-) []fantasy.Message {
-	if provider != fantasyanthropic.Name || len(messages) == 0 {
-		return messages
-	}
-
-	violations := chatprompt.ValidateAnthropicProviderToolHistory(messages)
-	if len(violations) == 0 {
-		return messages
-	}
-	affectedMessages := messageIndexesFromAnthropicProviderToolViolations(
-		violations,
-		len(messages),
-	)
-	guarded := sanitizeAnthropicProviderToolGuardMessages(
-		ctx,
-		logger,
-		provider,
-		modelName,
-		messages,
-		affectedMessages,
-		len(violations),
-	)
-	if isSafeAnthropicProviderToolPrompt(guarded) {
-		return guarded
-	}
-
-	fallbackViolations := chatprompt.ValidateAnthropicProviderToolHistory(guarded)
-	fallbackAffectedMessages := providerExecutedToolMessageIndexes(guarded)
-	guarded = sanitizeAnthropicProviderToolGuardMessages(
-		ctx,
-		logger,
-		provider,
-		modelName,
-		guarded,
-		fallbackAffectedMessages,
-		len(fallbackViolations),
-		slog.F("fallback", true),
-	)
-	if isSafeAnthropicProviderToolPrompt(guarded) {
-		return guarded
-	}
-
-	// The guard sanitizer should normally remove every typed provider block it
-	// selects. The strip path is a fail-closed backstop for analyzer and
-	// provider serialization drift, not a path we can drive without hooks.
-	preStripViolations := chatprompt.ValidateAnthropicProviderToolHistory(guarded)
-	stripMessages := messageIndexesFromAnthropicProviderToolViolations(
-		preStripViolations,
-		len(guarded),
-	)
-
-	var stripStats chatprompt.AnthropicProviderToolSanitizationStats
-	guarded, stripStats = stripAnthropicProviderToolHistoryFromMessages(
-		guarded,
-		stripMessages,
-	)
-	var sanitizeStats chatprompt.AnthropicProviderToolSanitizationStats
-	guarded, sanitizeStats = chatprompt.SanitizeAnthropicProviderToolHistory(
-		provider,
-		guarded,
-	)
-	stripStats = addAnthropicProviderToolSanitizationStats(stripStats, sanitizeStats)
-
-	if !isSafeAnthropicProviderToolPrompt(guarded) {
-		guarded, sanitizeStats = stripAnthropicProviderToolHistoryFromMessages(
-			guarded,
-			providerExecutedToolMessageIndexes(guarded),
-		)
-		stripStats = addAnthropicProviderToolSanitizationStats(stripStats, sanitizeStats)
-		guarded, sanitizeStats = chatprompt.SanitizeAnthropicProviderToolHistory(
-			provider,
-			guarded,
-		)
-		stripStats = addAnthropicProviderToolSanitizationStats(stripStats, sanitizeStats)
-		if !isSafeAnthropicProviderToolPrompt(guarded) {
-			logger.Error(
-				ctx,
-				"anthropic provider tool guard postcondition failed: prompt still unsafe after nuclear strip",
-				slog.F("phase", "pre_request_guard_postcondition_failed"),
-				slog.F("tool_type", "provider_executed"),
-				slog.F("provider", provider),
-				slog.F("model", modelName),
-			)
-		}
-	}
-
-	details, truncated := anthropicProviderToolViolationLogDetails(
-		preStripViolations,
-	)
-	chatprompt.LogAnthropicProviderToolSanitization(
-		ctx,
-		logger,
-		"pre_request_guard_fallback_strip",
-		provider,
-		modelName,
-		stripStats,
-		slog.F("validation_violations", len(preStripViolations)),
-		slog.F("validation_violation_details", details),
-		slog.F("truncated_violations", truncated),
-	)
-	return guarded
-}
-
-func sanitizeAnthropicProviderToolGuardMessages(
-	ctx context.Context,
-	logger slog.Logger,
-	provider string,
-	modelName string,
-	messages []fantasy.Message,
-	affectedMessages map[int]struct{},
-	validationViolations int,
-	extraFields ...slog.Field,
-) []fantasy.Message {
-	guardPrompt := invalidateProviderExecutedToolCallsInMessages(messages, affectedMessages)
-	// chatprompt owns message coalescing, empty message drops, and
-	// provider option preservation. Marking affected provider calls
-	// invalid lets its existing sanitizer remove the unsafe history.
-	sanitized, stats := chatprompt.SanitizeAnthropicProviderToolHistory(provider, guardPrompt)
-	extra := []slog.Field{
-		slog.F("validation_violations", validationViolations),
-	}
-	extra = append(extra, extraFields...)
-	chatprompt.LogAnthropicProviderToolSanitization(
-		ctx,
-		logger,
-		"pre_request_guard",
-		provider,
-		modelName,
-		stats,
-		extra...,
-	)
-	return sanitized
-}
-
-func isSafeAnthropicProviderToolPrompt(messages []fantasy.Message) bool {
-	return len(chatprompt.ValidateAnthropicProviderToolHistory(messages)) == 0
-}
-
-func messageIndexesFromAnthropicProviderToolViolations(
-	violations []chatprompt.AnthropicProviderToolHistoryViolation,
-	messageCount int,
-) map[int]struct{} {
-	indexes := make(map[int]struct{})
-	for _, violation := range violations {
-		if violation.MessageIndex < 0 || violation.MessageIndex >= messageCount {
-			continue
-		}
-		indexes[violation.MessageIndex] = struct{}{}
-	}
-	return indexes
-}
-
-func providerExecutedToolMessageIndexes(messages []fantasy.Message) map[int]struct{} {
-	indexes := make(map[int]struct{})
-	for messageIndex, message := range messages {
-		for _, part := range message.Content {
-			if toolCall, ok := safeToolCallPart(part); ok && toolCall.ProviderExecuted {
-				indexes[messageIndex] = struct{}{}
-				break
-			}
-			if toolResult, ok := safeToolResultPart(part); ok && toolResult.ProviderExecuted {
-				indexes[messageIndex] = struct{}{}
-				break
-			}
-		}
-	}
-	return indexes
-}
-
-func stripAnthropicProviderToolHistoryFromMessages(
-	messages []fantasy.Message,
-	affectedMessages map[int]struct{},
-) ([]fantasy.Message, chatprompt.AnthropicProviderToolSanitizationStats) {
-	var stats chatprompt.AnthropicProviderToolSanitizationStats
-	if len(affectedMessages) == 0 {
-		return messages, stats
-	}
-
-	out := make([]fantasy.Message, 0, len(messages))
-	for messageIndex, message := range messages {
-		if _, affected := affectedMessages[messageIndex]; !affected {
-			out = appendAnthropicProviderToolGuardMessage(out, message)
-			continue
-		}
-
-		parts := make([]fantasy.MessagePart, 0, len(message.Content))
-		for _, part := range message.Content {
-			if toolCall, ok := safeToolCallPart(part); ok && toolCall.ProviderExecuted {
-				stats.RemovedToolCalls++
-				continue
-			}
-			if toolResult, ok := safeToolResultPart(part); ok && toolResult.ProviderExecuted {
-				stats.RemovedToolResults++
-				if textPart, ok := chatprompt.AnthropicProviderToolResultTextPart(part); ok {
-					parts = append(parts, textPart)
-				}
-				continue
-			}
-			parts = append(parts, part)
-		}
-		if len(parts) == 0 {
-			stats.DroppedMessages++
-			continue
-		}
-		message.Content = parts
-		out = appendAnthropicProviderToolGuardMessage(out, message)
-	}
-	return out, stats
-}
-
-func appendAnthropicProviderToolGuardMessage(
-	out []fantasy.Message,
-	message fantasy.Message,
-) []fantasy.Message {
-	if len(out) == 0 || out[len(out)-1].Role != message.Role {
-		return append(out, message)
-	}
-
-	last := &out[len(out)-1]
-	lastContent := applyAnthropicProviderToolGuardOptionsToLastPart(
-		last.Content,
-		last.ProviderOptions,
-	)
-	messageContent := applyAnthropicProviderToolGuardOptionsToLastPart(
-		message.Content,
-		message.ProviderOptions,
-	)
-	content := make([]fantasy.MessagePart, 0, len(lastContent)+len(messageContent))
-	content = append(content, lastContent...)
-	content = append(content, messageContent...)
-	last.Content = content
-	last.ProviderOptions = nil
-	return out
-}
-
-func applyAnthropicProviderToolGuardOptionsToLastPart(
-	parts []fantasy.MessagePart,
-	options fantasy.ProviderOptions,
-) []fantasy.MessagePart {
-	if len(options) == 0 || len(parts) == 0 {
-		return parts
-	}
-
-	out := make([]fantasy.MessagePart, len(parts))
-	copy(out, parts)
-	lastIndex := len(out) - 1
-	switch part := out[lastIndex].(type) {
-	case fantasy.TextPart:
-		part.ProviderOptions = mergeAnthropicProviderToolGuardOptions(
-			part.ProviderOptions,
-			options,
-		)
-		out[lastIndex] = part
-	case *fantasy.TextPart:
-		if part != nil {
-			clone := *part
-			clone.ProviderOptions = mergeAnthropicProviderToolGuardOptions(
-				clone.ProviderOptions,
-				options,
-			)
-			out[lastIndex] = &clone
-		}
-	case fantasy.ReasoningPart:
-		part.ProviderOptions = mergeAnthropicProviderToolGuardOptions(
-			part.ProviderOptions,
-			options,
-		)
-		out[lastIndex] = part
-	case *fantasy.ReasoningPart:
-		if part != nil {
-			clone := *part
-			clone.ProviderOptions = mergeAnthropicProviderToolGuardOptions(
-				clone.ProviderOptions,
-				options,
-			)
-			out[lastIndex] = &clone
-		}
-	case fantasy.FilePart:
-		part.ProviderOptions = mergeAnthropicProviderToolGuardOptions(
-			part.ProviderOptions,
-			options,
-		)
-		out[lastIndex] = part
-	case *fantasy.FilePart:
-		if part != nil {
-			clone := *part
-			clone.ProviderOptions = mergeAnthropicProviderToolGuardOptions(
-				clone.ProviderOptions,
-				options,
-			)
-			out[lastIndex] = &clone
-		}
-	case fantasy.ToolCallPart:
-		part.ProviderOptions = mergeAnthropicProviderToolGuardOptions(
-			part.ProviderOptions,
-			options,
-		)
-		out[lastIndex] = part
-	case *fantasy.ToolCallPart:
-		if part != nil {
-			clone := *part
-			clone.ProviderOptions = mergeAnthropicProviderToolGuardOptions(
-				clone.ProviderOptions,
-				options,
-			)
-			out[lastIndex] = &clone
-		}
-	case fantasy.ToolResultPart:
-		part.ProviderOptions = mergeAnthropicProviderToolGuardOptions(
-			part.ProviderOptions,
-			options,
-		)
-		out[lastIndex] = part
-	case *fantasy.ToolResultPart:
-		if part != nil {
-			clone := *part
-			clone.ProviderOptions = mergeAnthropicProviderToolGuardOptions(
-				clone.ProviderOptions,
-				options,
-			)
-			out[lastIndex] = &clone
-		}
-	}
-	return out
-}
-
-func mergeAnthropicProviderToolGuardOptions(
-	first fantasy.ProviderOptions,
-	second fantasy.ProviderOptions,
-) fantasy.ProviderOptions {
-	if len(first) == 0 {
-		return second
-	}
-	if len(second) == 0 {
-		return first
-	}
-
-	merged := maps.Clone(first)
-	for provider, options := range second {
-		if options != nil {
-			merged[provider] = options
-		}
-	}
-	return merged
-}
-
-func addAnthropicProviderToolSanitizationStats(
-	first chatprompt.AnthropicProviderToolSanitizationStats,
-	second chatprompt.AnthropicProviderToolSanitizationStats,
-) chatprompt.AnthropicProviderToolSanitizationStats {
-	return chatprompt.AnthropicProviderToolSanitizationStats{
-		RemovedToolCalls:   first.RemovedToolCalls + second.RemovedToolCalls,
-		RemovedToolResults: first.RemovedToolResults + second.RemovedToolResults,
-		DroppedMessages:    first.DroppedMessages + second.DroppedMessages,
-	}
-}
-
-func anthropicProviderToolViolationLogDetails(
-	violations []chatprompt.AnthropicProviderToolHistoryViolation,
-) ([]map[string]any, bool) {
-	count := min(len(violations), maxAnthropicProviderToolViolationLogDetails)
-	details := make([]map[string]any, 0, count)
-	for _, violation := range violations[:count] {
-		details = append(details, map[string]any{
-			"message_index": violation.MessageIndex,
-			"part_index":    violation.PartIndex,
-			"id":            violation.ID,
-			"reason":        violation.Reason,
-		})
-	}
-	return details, len(violations) > maxAnthropicProviderToolViolationLogDetails
-}
-
-func invalidateProviderExecutedToolCallsInMessages(
-	messages []fantasy.Message,
-	affectedMessages map[int]struct{},
-) []fantasy.Message {
-	if len(affectedMessages) == 0 {
-		return messages
-	}
-	out := make([]fantasy.Message, len(messages))
-	copy(out, messages)
-	for messageIndex := range affectedMessages {
-		if messageIndex < 0 || messageIndex >= len(out) {
-			continue
-		}
-		message := out[messageIndex]
-		if len(message.Content) == 0 {
-			continue
-		}
-		parts := make([]fantasy.MessagePart, len(message.Content))
-		for partIndex, part := range message.Content {
-			parts[partIndex] = invalidateProviderExecutedToolCallPart(part)
-		}
-		message.Content = parts
-		out[messageIndex] = message
-	}
-	return out
-}
-
-func invalidateProviderExecutedToolCallPart(part fantasy.MessagePart) fantasy.MessagePart {
-	switch value := part.(type) {
-	case fantasy.ToolCallPart:
-		if value.ProviderExecuted {
-			value.ToolName = ""
-		}
-		return value
-	case *fantasy.ToolCallPart:
-		if value == nil {
-			return part
-		}
-		clone := *value
-		if clone.ProviderExecuted {
-			clone.ToolName = ""
-		}
-		return &clone
-	default:
-		return part
-	}
 }
 
 // guardedAttempt owns an attempt-scoped context and startup guard
@@ -1992,9 +1323,9 @@ func persistInterruptedStep(
 		provider = opts.Model.Provider()
 		modelName = opts.Model.Model()
 	}
-	var sanitizeStats chatprompt.AnthropicProviderToolSanitizationStats
-	result.content, sanitizeStats = sanitizeAnthropicProviderToolContent(provider, result.content)
-	chatprompt.LogAnthropicProviderToolSanitization(
+	var sanitizeStats chatsanitize.AnthropicProviderToolSanitizationStats
+	result.content, sanitizeStats = chatsanitize.SanitizeAnthropicProviderToolContent(provider, result.content)
+	chatsanitize.LogAnthropicProviderToolSanitization(
 		ctx, opts.Logger, "interrupted_persist", provider, modelName, sanitizeStats,
 	)
 
@@ -2032,7 +1363,7 @@ func persistInterruptedStep(
 		if _, exists := answeredToolCalls[tc.ToolCallID]; exists {
 			continue
 		}
-		if isAnthropicProviderExecutedToolCall(provider, tc) {
+		if chatsanitize.IsAnthropicProviderExecutedToolCall(provider, tc) {
 			continue
 		}
 		content = append(content, fantasy.ToolResultContent{
